Ticket 622 - DS logging errors "libdb: BDB0171 seek: 2147483648: (262144 * 8192) + 0: No such file or directory
Bug Description:  This error message gets logged in the error log if size of the id2entry grows beyond 2 GB.
                  The int storage class is the cause of the problem.

Fix description:  Use the proper storage class (off64_t)

https://fedorahosted.org/389/ticket/622

Reviewed by: richm(Thanks!)

+++ ldap/servers/slapd/back-ldbm/dblayer.c
@@ -510,7 +510,7 @@ int dblayer_open_huge_file(const char *path, int oflag, int mode)
 /* Helper function for large seeks, db4.3 */
 static int dblayer_seek43_large(int fd, off64_t offset, int whence)
 {
-    int ret = 0;
+    off64_t ret = 0;
 
     ret = lseek64(fd, offset, whence);
